SBK-RAM code refactor
