+++ sbk-api/src/main/java/io/sbk/perl/impl/TotalWindowLatencyRecorder.java
@@ -24,6 +24,6 @@ public class TotalWindowLatencyRecorder extends TotalLatencyRecordWindow impleme
     @Override
     public void recordLatency(long startTime, int bytes, int events, long latency) {
         window.recordLatency(startTime, bytes, events, latency);
-        windowOverflow(startTime);
+        checkWindowOverflowAndReset(startTime);
     }
 }
